+++ qiskit_nature/mapping/mappings/bravyi_kitaev_super_fast_mapping.py
@@ -10,18 +10,17 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""The Logarithmic Mapping interface."""
+"""The Super-Fast Bravyi-Kitaev Mapping interface."""
 
 from qiskit.aqua.operators import PauliSumOp
-
 from qiskit_nature.operators.second_quantization.particle_op import ParticleOp
-from qiskit_nature.operators.second_quantization.spin_op import SpinOp
+from qiskit_nature.operators.second_quantization.fermionic_op import FermionicOp
 
-from .qubit_mapping import QubitMapping
+from qiskit_nature.mapping.mappings.qubit_mapping import QubitMapping
 
 
-class LogarithmicMapping(QubitMapping):
-    """The Logarithmic spin-to-qubit mapping. """
+class BravyiKitaevSuperFastMapping(QubitMapping):
+    """The Super-Fast Bravyi-Kitaev fermion-to-qubit mapping. """
 
     def supports_particle_type(self, particle_type: ParticleOp) -> bool:
         """Returns whether the queried particle-type operator is supported by this mapping.
@@ -32,11 +31,11 @@ class LogarithmicMapping(QubitMapping):
         Returns:
             A boolean indicating whether the queried particle-type is supported.
         """
-        return isinstance(particle_type, SpinOp)
+        return isinstance(particle_type, FermionicOp)
 
     def map(self, second_q_op: ParticleOp) -> PauliSumOp:
-        """Maps a `ParticleOp` to a `PauliSumOp` using the Logarithmic spin-to-qubit
-        mapping.
+        """Maps a `ParticleOp` to a `PauliSumOp` using the Super-Fast Bravyi-Kitaev
+        fermion-to-qubit mapping.
 
         Args:
             second_q_op: the `ParticleOp` to be mapped.
