+++ qiskit_nature/mapping/mappings/parity_mapping.py
@@ -10,18 +10,18 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""The Direct Mapping interface."""
+"""The Parity Mapping interface."""
 
 from qiskit.aqua.operators import PauliSumOp
 
-from qiskit_nature.operators.second_quantization.bosonic_op import BosonicOp
+from qiskit_nature.operators.second_quantization.fermionic_op import FermionicOp
 from qiskit_nature.operators.second_quantization.particle_op import ParticleOp
 
-from .qubit_mapping import QubitMapping
+from qiskit_nature.mapping.mappings.qubit_mapping import QubitMapping
 
 
-class DirectMapping(QubitMapping):
-    """The Direct boson-to-qubit mapping. """
+class ParityMapping(QubitMapping):
+    """The Parity fermion-to-qubit mapping. """
 
     def supports_particle_type(self, particle_type: ParticleOp) -> bool:
         """Returns whether the queried particle-type operator is supported by this mapping.
@@ -32,10 +32,10 @@ class DirectMapping(QubitMapping):
         Returns:
             A boolean indicating whether the queried particle-type is supported.
         """
-        return isinstance(particle_type, BosonicOp)
+        return isinstance(particle_type, FermionicOp)
 
     def map(self, second_q_op: ParticleOp) -> PauliSumOp:
-        """Maps a `ParticleOp` to a `PauliSumOp` using the Direct boson-to-qubit
+        """Maps a `ParticleOp` to a `PauliSumOp` using the Parity fermion-to-qubit
         mapping.
 
         Args:
