+++ test/test_end2end_with_qpe.py
@@ -23,7 +23,7 @@ from qiskit.aqua import QuantumInstance
 from qiskit.aqua.algorithms import QPE, NumPyMinimumEigensolver
 from qiskit.aqua.operators import Z2Symmetries
 
-from qiskit_nature.mappings.mapped_ops_builder import mapping
+from qiskit_nature.mapping.mapped_ops_builder import mapping
 from test import QiskitNatureTestCase
 from qiskit_nature.drivers import PySCFDriver, UnitsType
 from qiskit_nature import FermionicOperator, QiskitNatureError
@@ -53,7 +53,7 @@ class TestEnd2EndWithQPE(QiskitNatureTestCase):
         fer_op = FermionicOperator(
             h1=molecule.one_body_integrals, h2=molecule.two_body_integrals)
         qubit_op = mapping(map_type=qubit_mapping, num_modes=fer_op.modes,
-                           h1=fer_op.h1, h2=fer_op.h2,
+                           h_1=fer_op.h1, h_2=fer_op.h2,
                            ph_trans_shift=fer_op._ph_trans_shift, threshold=1e-10)
         qubit_op = Z2Symmetries.two_qubit_reduction(qubit_op, 2)
